requirements.txt
pdfplumber
python-pptx
python-docx
pandas
sentence-transformers
faiss-cpu
numpy
fastapi
uvicorn
